services/ingestion/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
pydantic==2.8.2
pydantic-settings==2.4.0
aiokafka==0.10.0
influxdb-client==1.44.0
paho-mqtt==2.1.0
prometheus-client==0.20.0
python-json-logger==2.0.7
